LayoutTests/http/tests/security/contentSecurityPolicy/1.1/reflected-xss-allow-expected.txt

ALERT: XSS
Tests that 'X-WebKit-CSP: reflected-xss allow;' disables the XSSAuditor. This test passes if an alert is generated, and the script is allowed.
